+++ .babelrc
@@ -2,7 +2,6 @@
   "presets": ["@babel/preset-env"],
   "plugins": [
     "@babel/plugin-syntax-dynamic-import",
-    "@babel/transform-runtime",
-    "babel-plugin-rewire"
+    "@babel/transform-runtime"
   ]
 }
