+++ package.json
@@ -13,7 +13,7 @@
     "e2e": "node test/e2e/runner.js",
     "lint": "eslint --ext .js,.vue src --color",
     "npm-forever": "forever -o out.log -e err.log -c 'node --expose-gc' serverCluster.js",
-    "test": "jest tests/unit/store/mutations.test.js",
+    "test": "jest tests/unit/util/comm.test.js",
     "unit": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js --single-run",
     "unit-watch": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js",
     "validate": "npm ls"
@@ -100,7 +100,6 @@
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
     "babel-loader": "^8.0.4",
-    "babel-plugin-rewire": "^1.2.0",
     "chokidar": "^2.0.4",
     "eslint": "^5.8.0",
     "eslint-plugin-jest": "^22.0.0",
