+++ src/store/api.js
@@ -1,12 +1,14 @@
-import qs from 'qs'
+// this is aliased in webpack config based on server/client build
+import { CATEGORY, SECTION, TAG, TOPIC } from '../constants/index'
 import { camelizeKeys } from 'humps'
-import { get, merge } from 'lodash'
 import { getHost } from '../util/comm'
+import _ from 'lodash'
+import qs from 'qs'
 
 const superagent = require('superagent')
-const host = getHost()
+const _host = getHost()
 
-function buildQuery (params = {}) {
+function _buildQuery (params = {}) {
   let query = {}
   const whitelist = [ 'where', 'embedded', 'max_results', 'page', 'sort', 'related', 'clean', 'clientInfo', 'id', 'keyword', 'offset' ]
   whitelist.forEach((ele) => {
@@ -24,34 +26,35 @@ function buildQuery (params = {}) {
   return query
 }
 
-function buildUrlWithQuery ({ endpoint, params = {} }) {
-  let url = `${host}/api/${endpoint}`
-  const hasParams = params && Object.keys(params).length
-  if (hasParams) {
-    const query = buildQuery(params)
-    url = `${url}?${query}`
-  }
-  return url
+function _doFetch (url) {
+  return new Promise((resolve, reject) => {
+    superagent
+    .get(url)
+    // .query('')
+    .end(function (err, res) {
+      if (err) {
+        reject(err)
+      } else {
+        resolve(camelizeKeys(res.body))
+      }
+    })
+  })
 }
 
-function convertTypeForQuery (type) {
-  const typeList = {
-    SECTION: 'sections',
-    CATEGORY: 'categories',
-    TAG: 'tags',
-    TOPIC: 'topics'
-  }
-  return typeList[type]
+function _logClient (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/tracking`
+  url = `${url}?${query}`
+  return _doFetch(url)
 }
 
-function setupWhereInParam (type, value, params = {}) {
-  const convertedType = convertTypeForQuery(type)
+function _setupWhereInParam (key, value, params = {}) {
   params = params || {}
   value = Array.isArray(value) ? value : [ value ]
   const where = {}
-  if (convertedType && value.length > 0) {
-    merge(where, params.where, {
-      [convertedType]: {
+  if (value.length > 0) {
+    _.merge(where, params.where, {
+      [key]: {
         '$in': value
       }
     })
@@ -60,173 +63,350 @@ function setupWhereInParam (type, value, params = {}) {
   return params
 }
 
-async function doFetch (url) {
-  return await superagent
-    .get(url)
-    .then(res => camelizeKeys(res.body))
+function loadActivities (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/activities`
+  url = `${url}?${query}`
+  return _doFetch(url)
 }
 
-export const fetchActivities = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'activities', params })
-  return doFetch(url)
+function loadArticles (params = {}, preview) {
+  const query = _buildQuery(params)
+  let url = !preview ? `${_host}/api/posts` : `${_host}/api/drafts`
+  url = `${url}?${query}`
+  return _doFetch(url)
 }
 
-export const fetchArticles = (params = {}, preview) => {
-  const endpoint = !preview ? 'posts' : 'drafts'
-  const url = buildUrlWithQuery({ endpoint, params })
-  return doFetch(url)
+function loadArticlesGroupedList () {
+  const apiHost = `${_host}/api/grouped`
+  return _doFetch(apiHost)
 }
 
-export const fetchArticlesByUuid = (uuid = '', type = '', params = {}) => {
-  params = setupWhereInParam(type, [ uuid ], params)
+function loadArticlesPopList () {
+  const url = `${_host}/api/poplist`
+  return _doFetch(url)
+}
+
+function loadArticlesByUuid (uuid = '', type = '', params = {}) {
+  switch (type) {
+    case SECTION:
+      params = _setupWhereInParam('sections', [ uuid ], params)
+      break
+    case CATEGORY:
+      params = _setupWhereInParam('categories', [ uuid ], params)
+      break
+    case TAG:
+      params = _setupWhereInParam('tags', [ uuid ], params)
+      break
+    case TOPIC:
+      params = _setupWhereInParam('topics', [ uuid ], params)
+      break
+    default:
+      return Promise.resolve()
+  }
   params.related ? (params.useMetaEndpoint = true) : ''
   params.sort = params.sort || '-publishedDate'
-  const endpoint = params.useMetaEndpoint ? 'meta' : 'listing'
-  const url = buildUrlWithQuery({ endpoint, params })
-  return doFetch(url)
+  let url
+  params.useMetaEndpoint ? url = `${_host}/api/meta` : url = `${_host}/api/listing`
+  const query = _buildQuery(params)
+  url = `${url}?${query}`
+  return _doFetch(url)
 }
 
-export const fetchArticlesGroupedList = () => {
-  const apiHost = `${host}/api/grouped`
-  return doFetch(apiHost)
+function loadAudios (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/audios`
+  url = `${url}?${query}`
+  return _doFetch(url)
 }
 
-export const fetchArticlesPopList = () => {
-  const url = `${host}/api/poplist`
-  return doFetch(url)
-}
-
-export const fetchAudios = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'audios', params })
-  return doFetch(url)
-}
-
-export const fetchCommonData = (endpoints = []) => {
-  const comboParams = endpoints.map(endpoint => `endpoint=${endpoint}`).join('&')
-  let url = `${host}/api/combo?`
+function loadCommonData (endpoints = []) {
+  const mapped = _.map(endpoints, (n) => { return 'endpoint=' + n })
+  const comboParams = mapped.join('&')
+  let url = `${_host}/api/combo?`
   url = url + comboParams
-  return doFetch(url).then(data => {
-    Object.keys(get(data, 'endpoints', {})).map(endpoint => {
-      if (endpoint === 'sections') { // get categories data from sections
-        data.endpoints.categories = {}
-        data.endpoints[endpoint].items.map(section => {
-          section.categories.filter(category => category.name).map(category => {
-            data.endpoints.categories[category.name] = category
+  return _doFetch(url)
+}
+
+function loadContacts (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/contacts`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadData (params = {}, endpoint) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/${endpoint}`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadEditorChoice () {
+  const url = `${_host}/api/combo?endpoint=choices`
+  return _doFetch(url)
+}
+
+function loadEvent (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/event`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadExternals (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/externals`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadImages (uuid = '', type = '', params = {}) {
+  switch (type) {
+    case SECTION:
+      params = _setupWhereInParam('sections', [ uuid ], params)
+      break
+    case CATEGORY:
+      params = _setupWhereInParam('categories', [ uuid ], params)
+      break
+    case TAG:
+      params = _setupWhereInParam('tags', [ uuid ], params)
+      break
+    case TOPIC:
+      params = _setupWhereInParam('topics', [ uuid ], params)
+      break
+    default:
+      return Promise.resolve()
+  }
+
+  const query = _buildQuery(params)
+  let url = `${_host}/api/images`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadImagesById (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/images`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadImage (uuid = '') {
+  const url = `${_host}/api/images/${uuid}`
+  return _doFetch(url)
+}
+
+function loadLatestArticle (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/listing`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadLatestNewsFromJson () {
+  const url = `${_host}/api/latestNews`
+  return _doFetch(url)
+}
+
+function loadNodes (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/nodes`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadOathPlaylist ({ id, params }) {
+  let url = `${_host}/api/playlistng/${id}`
+  if (params) {
+    const query = _buildQuery(params)
+    url = `${url}?${query}`
+  }
+  return _doFetch(url)
+}
+
+
+function loadOathVideo ({ id, params }) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/video/${id}`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadOathVideoByPlaylist ({ id, params }) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/video/playlist/${id}`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadPartners (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/partners`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadSearch (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/search`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadSectionList () {
+  const url = `${_host}/api/combo?endpoint=sections`
+  return _doFetch(url)
+}
+
+function loadTag (slug = '') {
+  const url = `${_host}/api/tags/${slug}`
+  return _doFetch(url)
+}
+
+function loadTimeline (slug = '') {
+  const url = `${_host}/api/timeline/${slug}`
+  return _doFetch(url)
+}
+
+function loadTopic (params = {}) {
+  const query = _buildQuery(params)
+  let url = `${_host}/api/topics`
+  url = `${url}?${query}`
+  return _doFetch(url)
+}
+
+function loadYoutubePlaylist (limit = 12, pageToken = '') {
+  const url = `${_host}/api/playlist?maxResults=${limit}&pageToken=${pageToken}`
+  return _doFetch(url)
+}
+
+export function logClient (params = {}) {
+  return _logClient(params)
+}
+
+export function fetchActivities (params = {}) {
+  return loadActivities(params)
+}
+
+export function fetchArticles (params = {}, preview) {
+  return loadArticles(params, preview)
+}
+
+export function fetchArticlesByUuid (uuid = '', type = '', params = {}, isOnlyMeta = true) {
+  return loadArticlesByUuid(uuid, type, params, isOnlyMeta)
+}
+
+export function fetchArticlesGroupedList (params = {}) {
+  return loadArticlesGroupedList(params)
+}
+
+export function fetchArticlesPopList (params = {}) {
+  return loadArticlesPopList(params)
+}
+
+export function fetchAudios (params = {}) {
+  return loadAudios(params)
+}
+
+export function fetchCommonData (endpoints = []) {
+  return Promise.all([ loadCommonData(endpoints) ])
+  .then((data) => {
+    const commonData = {}
+    _.map(Object.keys(_.get(data[0], [ 'endpoints' ])), (e) => {
+      commonData[e] = _.get(data[0], [ 'endpoints', e ])
+      if (e === 'sections') {
+        _.forEach(_.get(data[0], [ 'endpoints', e, 'items' ]), (s) => {
+          _.forEach(s.categories, (c) => {
+            _.set(commonData, [ 'categories', c.name ], c)
           })
         })
       }
     })
-    return data.endpoints
+    return commonData
   })
 }
 
-export const fetchContacts = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'contacts', params })
-  return doFetch(url)
+export function fetchContacts (params = {}) {
+  return loadContacts(params)
 }
 
-export const fetchEditorChoice = () => {
-  const url = `${host}/api/combo?endpoint=choices`
-  return doFetch(url)
+export function fetchEditorChoice () {
+  return loadEditorChoice()
 }
 
-export const fetchEvent = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'event', params })
-  return doFetch(url)
+export function fetchEvent (params = {}) {
+  return loadEvent(params)
 }
 
-export const fetchExternals = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'externals', params })
-  return doFetch(url)
+export function fetchExternals (params = {}) {
+  return loadExternals(params)
 }
 
-export const fetchImage = (uuid = '') => {
-  const url = `${host}/api/images/${uuid}`
-  return doFetch(url)
+export function fetchImage (uuid = '') {
+  return loadImage(uuid)
 }
 
-export const fetchImages = (uuid = '', type = '', params = {}) => {
-  params = setupWhereInParam(type, [ uuid ], params)
-  const url = buildUrlWithQuery({ endpoint: 'images', params })
-  return doFetch(url)
+export function fetchImages (uuid = '', type = '', params = {}) {
+  return loadImages(uuid, type, params)
 }
 
-export const fetchImagesById = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'images', params })
-  return doFetch(url)
+export function fetchImagesById (params = {}) {
+  return loadImagesById(params)
 }
 
-export const fetchLatestArticle = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'listing', params })
-  return doFetch(url)
+export function fetchLatestArticle (params = {}) {
+  return loadLatestArticle(params)
 }
 
-export const fetchLatestNewsFromJson = () => {
-  const url = `${host}/api/latestNews`
-  return doFetch(url)
+export function fetchLatestNewsFromJson () {
+  return loadLatestNewsFromJson()
 }
 
-export const fetchNodes = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'nodes', params })
-  return doFetch(url)
+export function fetchNodes (params = {}) {
+  return loadNodes(params)
 }
 
-export const fetchOathPlaylist = ({ id = '', params }) => {
-  const url = buildUrlWithQuery({ endpoint: `playlistng/${id}`, params })
-  return doFetch(url)
+export function fetchOathPlaylist ({ id = '', params }) {
+  return loadOathPlaylist({ id, params })
 }
 
-export const fetchOathVideo = ({ id = '', params = {} }) => {
-  const url = buildUrlWithQuery({ endpoint: `video/${id}`, params })
-  return doFetch(url)
+export function fetchOathVideo ({ id = '', params = {} }) {
+  return loadOathVideo({ id, params })
 }
 
-export const fetchOathVideoByPlaylist = ({ id = '', params = {} }) => {
-  const url = buildUrlWithQuery({ endpoint: `video/playlist/${id}`, params })
-  return doFetch(url)
+export function fetchOathVideoByPlaylist ({ id = '', params = {} }) {
+  return loadOathVideoByPlaylist({ id, params })
 }
 
-export const fetchPartners = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'partners', params })
-  return doFetch(url)
+export function fetchPartners (params = {}) {
+  return loadPartners(params)
 }
 
-export const fetchRecommendList = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'related_news', params })
-  return doFetch(url)
+export function fetchRecommendList (params = {}) {
+  return loadData(params, 'related_news')
 }
 
-export const fetchSearch = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'search', params })
-  return doFetch(url)
+export function fetchSearch (params = {}) {
+  return loadSearch(params)
 }
 
-export const fetchSectionList = () => {
-  const url = `${host}/api/combo?endpoint=sections`
-  return doFetch(url)
+export function fetchSectionList () {
+  return loadSectionList()
 }
 
-export const fetchTag = (slug = '') => {
-  const url = `${host}/api/tags/${slug}`
-  return doFetch(url)
+export function fetchTag (slug = '') {
+  return loadTag(slug)
 }
 
-export const fetchTimeline = (slug = '') => {
-  const url = `${host}/api/timeline/${slug}`
-  return doFetch(url)
+export function fetchTimeline (slug = '') {
+  return loadTimeline(slug)
 }
 
-export const fetchTopic = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'topics', params })
-  return doFetch(url)
+export function fetchTopic (params = {}) {
+  return loadTopic(params)
 }
 
-export const fetchYoutubePlaylist = (limit = 12, pageToken = '') => {
-  const url = `${host}/api/playlist?maxResults=${limit}&pageToken=${pageToken}`
-  return doFetch(url)
-}
-
-export const logClient = (params = {}) => {
-  const url = buildUrlWithQuery({ endpoint: 'tracking', params })
-  return doFetch(url)
+export function fetchYoutubePlaylist (limit = {}, pageToken = '') {
+  return loadYoutubePlaylist(limit, pageToken)
 }
